update: export module (#222)
