+++ server-v2/api/studio/internal/svc/servicecontext.go
@@ -8,9 +8,9 @@ import (
 	"github.com/vesoft-inc/go-pkg/httpclient"
 	"github.com/vesoft-inc/go-pkg/response"
 	"github.com/vesoft-inc/go-pkg/validator"
-	"github.com/vesoft-inc/nebula-studio/server-v2/api/studio/internal/common"
 	"github.com/vesoft-inc/nebula-studio/server-v2/api/studio/internal/config"
 	"github.com/vesoft-inc/nebula-studio/server-v2/api/studio/pkg/ecode"
+	"github.com/vesoft-inc/nebula-studio/server-v2/api/studio/pkg/utils"
 	"github.com/zeromicro/go-zero/core/logx"
 )
 
@@ -51,7 +51,7 @@ func createResponseHandler(c config.Config) response.Handler { // nolint:gocriti
 		Errorf:    logx.Errorf,
 		DebugInfo: c.Debug.Enable,
 		CheckBodyType: func(r *http.Request) response.StandardHandlerBodyType {
-			if common.PathMatchPattern(r.URL.Path, common.IgnoreHandlerBodyPatterns) {
+			if utils.PathMatchPattern(r.URL.Path, utils.IgnoreHandlerBodyPatterns) {
 				return response.StandardHandlerBodyNone
 			}
 			return response.StandardHandlerBodyJson
