+++ server-v2/api/studio/pkg/server/server.go
@@ -1,0 +1,18 @@
+package server
+
+import (
+	"github.com/vesoft-inc/nebula-studio/server-v2/api/studio/internal/config"
+	"github.com/vesoft-inc/nebula-studio/server-v2/api/studio/internal/service/importer"
+	"github.com/vesoft-inc/nebula-studio/server-v2/api/studio/internal/svc"
+)
+
+type (
+	ServiceContext = svc.ServiceContext
+	Config         = config.Config
+)
+
+var NewServiceContext = svc.NewServiceContext
+
+func InitDB(dbFilePath string) {
+	importer.InitDB(dbFilePath)
+}
